requirements.txt
kafka-python
zookeeper
river
pyyaml
path
